OS/2: QProcess: Don't close the stdin pipe until the child reads it all.
This fixes the cases when the child process receives truncated data sent to its stdin by the parent process with QProcess::write(). This is due to the fact that writing a chunk of data to a pipe and then closing it will make this data unavailable to the other end if it is not fast enough to read it in between.