OS/2: QProcess: Fixed the way how pipes are closed.
This makes sure that feof() will return 1 and ferror() will not
return 1 on the child end of the pipe connected to its stdin.
This in turn fixes programs that check the ferror() value to
decide if the input data is valid, as well as programs using
feof() as the only way to break the input stream read loop.