QFile: Fixed open() docs.
The comment "Sequential devices, such as stdin and stdout..." in
the Windows section is true only partly -- as long as these devices are
attached to the console which performs all kinds of CRLF normalization
on its own. If they are redirected to regular files, then this normalzation
is not done which leads to all kinds of weird behavior (duplicate CR
symbols on output, hangs due to atEnd() returning false on input).

The affected warning is also valid on OS/2 so it now mentions this.