uic: Put stdin/stdout to binary mode on Windows and OS/2.
This is necessary to avoid duplication of CR symbols on output
and hangs in atEnd() due to unexpected CR on input (in particular,
seen when stdin and stdout are redirected to regular files). See
also r932.

Note: it actually may make sense to change QFile::open() so that
it implicitly makes this mode change and restores the origianl mode
on destruction. Just to be on the safe side.