OS/2: xsystray: Fixed duplicate mouse click notifications.
This was due to the fact that the pointer to the notification data
passed to the client was not fixed up after making a local copy of
this data. As a result it would still point to the shared data cell
that could get already overwritten with a different notificaiton
message by the time it got processed in the client.