qmake: gnumake/os2: Make .sym generation part of main target rule.
Previously the .sym file was a separate rule depending on the target.
However, it actually needed not the target itself, but a map file generated
when generating the target. Since for EXE targets different builds
generate the EXE file under the same name but may not generate the
map file, the .sym generation could fail (e.g. if building the release after
the debug).

Note that under the new scheme there is a problem of the .sym file not
being generated at all in this case. But that's a different story since having
two distinct targets (release and debug) output to the same file is
confusing per se and there is no clean solution for that other than using
separate dirs for debug/release (which is always a good thing).