qmake: Ensure Makefile generation order in ordered builds.
When one sub-project links to another sub-project's library, ordered
builds are used. However, it's not enough only to build the projects in
the respective order, Makefile generation should be ordered too. One
reason for that is projects using library names with version suffixes
which qmake deduces at Makefile generation time from the .prl file if
the actual library file is not there (not yet built). The .prl file is
generated along with the respective Makefile; hence the dependency.

This applicable to at least Windows and OS/2 and fixes parallel build issues
with some libraries (e.g. QtHelp which depends on QtCLucene).