mkspecs/os2-g++: Always add project directory to INCLUDEPATH.
This is what qmake-based applications expect. On Unix, this is done
right in the respective qmake generator. On Windows, the compiler seems
to automatically add the directory of the file being compiled to the list of
includes, so no need to do that explicitly. More details are in
#227:comment:7.

This change makes it unnecessary to add "INCLUDEPATH = ." to .pro files in
order to fix build breaks we were sometimes suffering from on OS/2.