OS/2: configure.cmd: Remove qmake from default build parts.
In fact, when doing a parallel build (with make -jN), it is possible (even
after r976) that linking qmake.exe may happen at the same time as
generating a Makefile for other sub-target (i.e. when qmake.exe is in use
and cannot be overwritten). Given that building qmake from .pro is
interesting only when debugging qmake itself (so quite rarely), it doesn't
make much sense to fix this problem; removing qmake from the grand
build process is enough. It can alwasy be built manually if needed.