OS/2: Create native widget siblings and children when native ID is requested.
This makes sure each sibling and child of the widget for which a native
window is created will get a native window too. This is necessary to let
the system do proper clipping when the widget is painted on externally
using its native window ID.

Note that on OS/2 system clipping is not done (see QWidget::create_sys())
so qt_WinProcessWindowObstacles() needs to be used to clip out areas
obscured by other widgets when directly painting on the widget bypassing
Qt.