OS/2: Fix position errors on resize of child widgets.
During complex geometry change operations (e.g. directly resizing
the child widget managed by the layout manager of its parent and
then resizing the parent) the Y position of the child within the parent
could be shifted by the double delta of the resize value because
the new parent's height (in Qt terms) was used for flipping the Y
coordinate instead of the real height of the PM window which didn't
yet reflect the height change in Qt.