OS/2: Take style icons for SP_FileIcon & SP_DirIcon from QFileIconProvider.
This makes sure that the icons are consistent (taken from the system theme)
both in QCommonStyle and in QFileIconProvider. Previously, the former would
return the stock Qt icons which are completely different which made the 
Projects view in Qt Creator look odd compared to the FileSystem view.

Note that this (as many other similar settings as well) is only done when
QApplication::desktopSettingsAware() is true.