OS/2: Remove suport for system-wide qt.conf files.
This turns not to be necessary. The new code is much simpler and has
less platform-dependent details. Note that hard-coded component paths
are now relative to the parent of the directory containing QtCore4.dll, not
to that directory itself as before. This in particular makes qt.conf not necessary
for the development builds at all.

For official (RPM) release builds, full hard-coded paths will be used so qt.conf
is not necessary at all. It will only be in use in portable ZIP distributions to
override these hard-coded paths of RPM builds.