OS/2: Fix native child widget painting in DIVE mode.
The dirty regions of nested child widgets were calculated incorrectly since the
HWND of the top-level window was used instead of the child HWND.

Note that this didn't show problems in Qt 4.6.3 since requesting a native HWND
for a child (which is used by video players to embed video to the Qt
application) didn't cause its siblings and sibligs of its parents to get native
HWNDs and therefore they didn't use the code that contained this bug. But in
r1012 this incorrect behavior was fixed and the problem immediately popped up.