Don't eat drive letters in QDir::cleanPath().
For paths like "D:/../myfile", it would return "myfile" on all platforms while on
e.g. Windows and OS/2 "D:/" means the root directory and must not be removed.
Now, cleanPath() will leave it as "D:/../myfile" in such cases (similar to Linux
returning "/../myfile" unmodified).

The correctness of "/../myfile" itself is questionable (at least Linux treats such
paths as if there was no ".." and so does the native OS/2 call when resolving the
absolute file path), but it's a different story.