Assistant: Fix signal recursion at startup in remote control mode.
The index update process was restarting once finished until both the
contents and the index would happen to be created at the same time.
This caused a significant delay at startup (this process could never end,
actually) accompanied by the 100% CPU load during which remote
commands could not be processed.