syncqt: Paths in headers.pri must be relative to QTDIR.
These paths are used in .pro files located in the Qt source tree (QTDIR)
so they must be relative to it in order to get correct final paths (not to
the build directory which differs in shadow builds).

The code only worked with shadow builds only if the shadow build tree
was located in the same parent directory as the Qt source tree (i.e.
/home/bob/Coding/qt and /home/bob/Coding/qt-build). This is not always
the case.