corelib: Make QFileInfo returned by QDirIterator report symlinks properly.
In QT_OS2_USE_DOSFINDFIRST mode (which is ON now), we use
DosFindFirst which is unaware of symlinks and doesn't properly resolve that.
Given that in many other places kLIBC functions are used which handle symlinks
pretty well, this would completely confuse some apps like QFileDialog (see #284).

This fix removes this inconsistency.